requirements.txt
git+https://github.com/Sayvai-io/custom-tools.git
fastapi==0.103.1
langchain==0.0.292
pinecone_client==2.2.2
pydantic==1.10.11
SQLAlchemy==2.0.20
